+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+##[Unreleased]
+
+### Added
+- `status_bar.fullscreen_visible` setting to allow status bar to be visible in fullscreen  
+- `tablist.fullscreen_visible` setting to allow the tablist to be visible in fullscreen  
+
 ## [2.4.0]
 
 ### Added